+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 
 # Development
+## 0.0.3
+`+` Add link for favicon, no icon yet  
+`+` Add a bunch of logging  
+`*` Overhaul to use `Koa` (and HTTP/2) over `Express`  
+
 ## 0.0.2
 `+` New grunt tasks (TS, SASS, Babel)  
 `+` Create entire React app with esmodules